Some refactoring to allow cli calls from other python scripts
