+++ docker/.env.example
@@ -1,6 +1,13 @@
 URL=https://digicampus.uni-augsburg.de
 USER=<user>
 PASS=<password>
+# Specify which courses should be downloaded.
+# Possible values:
+#  - ALL
+#  - CURRENT
+#  - list in the form of: "<semester name>\n<course name>\n<semester name>\n<course name>\n..."
+DOWNLOAD_COURSES=CURRENT
+DOWNLOAD_LOCATION=/download
 
 # Optional regex for filtering downloaded files
 #REGEX=.+\..+
@@ -12,5 +19,5 @@ FILES_GROUP=33
 # Nextcloud specific configuration:
 #NEXTCLOUD_HOSTNAME=<ip or hostname>
 
-# Originates in naxtcloud data folder
+# Originates in nextcloud data folder
 #NEXTCLOUD_SCAN_FOLDER=<username>/files/<path in user files>
