+++ requirements.txt
@@ -2,3 +2,4 @@ BeautifulSoup4
 lxml
 requests
 clint
+python-dotenv
